Tree: dedup cursor logic
